refactoring PR comments
